refactor app and make numerous updates and bug fixes
